+++ chopper_generator/CHANGELOG.md
@@ -1,8 +1,27 @@
 # Changelog
 
+## 2.5.0
+
+- Unsuccessful response are not throw anymore, use `Response.isSuccessful` getter or `statusCode` instead
+- Support HEAD request
+
+## 2.4.2
+
+- Fix on JsonConverter 
+    If content type header overrided using @Post(headers: {'content-type': '...'})
+    The converter won't add json header and won't apply json.encode if content type is not JSON
+
+- add `bool override` on `applyHeader(s)` functions, true by default
+
+- support `List<MultipartFile>`
+
+## 2.4.1
+
+- Fix PartValue cast
+
 ## 2.4.0
 
-- Deprecate `@FieldField`, use `@PartFile` instead
+- Deprecate `@FileField`, use `@PartFile` instead
 - support `chopper: ^2.4.1`
 
 ## 2.3.4